requirements.txt
pandas==2.1.1
numpy==1.26.2
pyarrow==12.0.0
scipy==1.11.3
matplotlib==3.8.1
pytest==7.4.0
pyyaml==6.0
